requirements.txt
youtube-transcript-api
transformers
sentencepiece
torch
openai-whisper
pydub
streamlit
flask
